datapath/fake: Move BandwidthManager to fake/types package
See previous commit for rationale.
